pyproject.toml
[tool.poetry]
name = "place"
version = "0.1.0"
description = ""
authors = ["嘉然今天吃什么 <jsclndnz@outlook.com>"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.75.1"}
tortoise-orm = {extras = ["accel"], version = "^0.18.1"}
aiomysql = "^0.0.22"
aiosqlite = "^0.17.0"
pytz = "^2022.1"
python-dateutil = "^2.8.2"
aerich = "^0.6.2"
cryptography = "^36.0.2"
Pillow = "^9.1.0"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.18.1"
pytest-xdist = "^2.5.0"
pytest-ordering = "^0.6"

[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
